+++ @vates/types/src/index.mts
@@ -6,4 +6,3 @@ export * from './lib/xen-orchestra-xapi.mjs'
 export * from './lib/xen-orchestra-acl.mjs'
 export * from './lib/complex-matcher.mjs'
 export * from './lib/rest-api.mjs'
-export * from './lib/vhd-lib.mjs'
