+++ @xen-orchestra/disk-cli/package.json
@@ -20,7 +20,6 @@
   },
   "dependencies": {
     "@vates/async-each": "^1.0.3",
-    "@vates/types": "^2.0.1",
     "@xen-orchestra/backups": "^0.74.2",
     "@xen-orchestra/backup-archive": "^1.1.2",
     "@xen-orchestra/disk-transform": "^1.3.3",
